lesson_10/requirements.txt
selenium>=4.15.0
pytest>=7.4.0
allure-pytest>=2.13.0
pytest-xdist>=3.3.0
